requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-core>=0.1.0

# LangGraph for workflow orchestration
langgraph>=0.0.15

# Web application
streamlit>=1.28.0

# Additional dependencies
pydantic>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0